agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=1.0.0
boto3>=1.35.0
